moved sidebar svg to utils folder, remove unused components from superadmin
