+++ frontend/src/app/superadmin/layout.tsx
@@ -3,6 +3,7 @@ import React, { useState, useEffect } from "react";
 import { Sidebar } from "@/components/common/Sidebar";
 import { HeaderSection } from "@/components/superadmin/HeaderSection";
 import { usePathname, useRouter } from "next/navigation";
+import { DashboardIcon, UsersIcon, SchoolsIcon } from "@/utils/icons";
 
 export const Layout = ({ children }: {children: React.ReactNode}) => {
   const router = useRouter();
@@ -14,15 +15,15 @@ export const Layout = ({ children }: {children: React.ReactNode}) => {
 
   const sidebarItems = [
     {
-      icon: '<svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M10.8333 6.66667V3.33333C10.8333 3.09722 10.9133 2.89944 11.0733 2.74C11.2333 2.58056 11.4311 2.50056 11.6667 2.5H16.6667C16.9028 2.5 17.1008 2.58 17.2608 2.74C17.4208 2.9 17.5006 3.09778 17.5 3.33333V6.66667C17.5 6.90278 17.42 7.10083 17.26 7.26083C17.1 7.42083 16.9022 7.50056 16.6667 7.5H11.6667C11.4306 7.5 11.2328 7.42 11.0733 7.26C10.9139 7.1 10.8339 6.90222 10.8333 6.66667ZM2.5 10V3.33333C2.5 3.09722 2.58 2.89944 2.74 2.74C2.9 2.58056 3.09778 2.50056 3.33333 2.5H8.33333C8.56944 2.5 8.7675 2.58 8.9275 2.74C9.0875 2.9 9.16722 3.09778 9.16667 3.33333V10C9.16667 10.2361 9.08667 10.4342 8.92667 10.5942C8.76667 10.7542 8.56889 10.8339 8.33333 10.8333H3.33333C3.09722 10.8333 2.89944 10.7533 2.74 10.5933C2.58056 10.4333 2.50056 10.2356 2.5 10ZM10.8333 16.6667V10C10.8333 9.76389 10.9133 9.56611 11.0733 9.40667C11.2333 9.24722 11.4311 9.16722 11.6667 9.16667H16.6667C16.9028 9.16667 17.1008 9.24667 17.2608 9.40667C17.4208 9.56667 17.5006 9.76444 17.5 10V16.6667C17.5 16.9028 17.42 17.1008 17.26 17.2608C17.1 17.4208 16.9022 17.5006 16.6667 17.5H11.6667C11.4306 17.5 11.2328 17.42 11.0733 17.26C10.9139 17.1 10.8339 16.9022 10.8333 16.6667ZM2.5 16.6667V13.3333C2.5 13.0972 2.58 12.8994 2.74 12.74C2.9 12.5806 3.09778 12.5006 3.33333 12.5H8.33333C8.56944 12.5 8.7675 12.58 8.9275 12.74C9.0875 12.9 9.16722 13.0978 9.16667 13.3333V16.6667C9.16667 16.9028 9.08667 17.1008 8.92667 17.2608C8.76667 17.4208 8.56889 17.5006 8.33333 17.5H3.33333C3.09722 17.5 2.89944 17.42 2.74 17.26C2.58056 17.1 2.50056 16.9022 2.5 16.6667ZM4.16667 9.16667H7.5V4.16667H4.16667V9.16667ZM12.5 15.8333H15.8333V10.8333H12.5V15.8333ZM12.5 5.83333H15.8333V4.16667H12.5V5.83333ZM4.16667 15.8333H7.5V14.1667H4.16667V15.8333Z" fill="#5B5B5B"></path></svg>',
+      icon: DashboardIcon,
       label: "Dashboard",
     },
     {
-      icon: '<svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M13.3332 14.1669V15.8335H1.6665V14.1669C1.6665 14.1669 1.6665 10.8335 7.49984 10.8335C13.3332 10.8335 13.3332 14.1669 13.3332 14.1669ZM10.4165 6.25019C10.4165 5.67333 10.2454 5.10942 9.92496 4.62978C9.60447 4.15014 9.14895 3.7763 8.616 3.55554C8.08305 3.33479 7.4966 3.27703 6.93082 3.38957C6.36505 3.50211 5.84535 3.7799 5.43744 4.1878C5.02954 4.5957 4.75175 5.1154 4.63921 5.68118C4.52667 6.24696 4.58443 6.8334 4.80519 7.36635C5.02594 7.8993 5.39978 8.35483 5.87942 8.67531C6.35907 8.9958 6.92298 9.16686 7.49984 9.16686C8.27339 9.16686 9.01525 8.85957 9.56223 8.31259C10.1092 7.76561 10.4165 7.02374 10.4165 6.25019ZM13.2832 10.8335C13.7955 11.23 14.2146 11.7339 14.5112 12.3098C14.8078 12.8857 14.9746 13.5196 14.9998 14.1669V15.8335H18.3332V14.1669C18.3332 14.1669 18.3332 11.1419 13.2832 10.8335ZM12.4998 3.33353C11.9263 3.33086 11.3655 3.5023 10.8915 3.82519C11.3977 4.53247 11.6699 5.38043 11.6699 6.25019C11.6699 7.11995 11.3977 7.96791 10.8915 8.67519C11.3655 8.99808 11.9263 9.16953 12.4998 9.16686C13.2734 9.16686 14.0153 8.85957 14.5622 8.31259C15.1092 7.76561 15.4165 7.02374 15.4165 6.25019C15.4165 5.47665 15.1092 4.73478 14.5622 4.1878C14.0153 3.64082 13.2734 3.33353 12.4998 3.33353Z" fill="white"></path></svg>',
+      icon: UsersIcon,      
       label: "Users",
     },
     {
-      icon: '<svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M10.0002 0.488281L15.0002 5.48828V7.49995H18.3335V15.8333H19.1668V17.4999H0.833496V15.8333H1.66683V7.49995H5.00016V5.48828L10.0002 0.488281ZM15.0002 15.8333H16.6668V9.16661H15.0002V15.8333ZM5.00016 9.16661H3.3335V15.8333H5.00016V9.16661ZM6.66683 6.17828V15.8333H9.16683V9.99995H10.8335V15.8333H13.3335V6.17828L10.0002 2.84495L6.66683 6.17828Z" fill="#5B5B5B"></path></svg>',
+      icon: SchoolsIcon,      
       label: "Schools",
     },
   ];
