--- frontend/src/components/superadmin/users/icons.tsx
@@ -1,37 +1,0 @@
-import React from "react";
-
-export const CloseIcon: React.FC = () => (
-  <div
-    dangerouslySetInnerHTML={{
-      __html:
-        '<svg id="52:3481" layer-name="Close" width="12" height="12" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg" class="close-icon" style="width: 10px; height: 10px; stroke-width: 1.5px; stroke: #5B5B5B"> <path d="M11 1.00004L1 11M0.999958 1L10.9999 11" stroke="#5B5B5B" stroke-width="1.5" stroke-linecap="round"></path> </svg>',
-    }}
-  />
-);
-
-export const DropdownIcon: React.FC = () => (
-  <div
-    dangerouslySetInnerHTML={{
-      __html:
-        '<svg id="52:3440" layer-name="mdi:arrow-down-drop" width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg" class="dropdown-icon" style="width: 16px; height: 16px; margin-left: auto"> <path d="M3 6L8 11L13 6H3Z" fill="#464646"></path> </svg>',
-    }}
-  />
-);
-
-export const DropdownUpIcon: React.FC = () => (
-  <div
-    dangerouslySetInnerHTML={{
-      __html:
-        '<svg id="52:3476" layer-name="mdi:arrow-down-drop" width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg" class="dropdown-icon" style="width: 16px; height: 16px; margin-left: auto"> <path d="M13 10L8 5L3 10L13 10Z" fill="#464646"></path> </svg>',
-    }}
-  />
-);
-
-export const CheckIcon: React.FC = () => (
-  <div
-    dangerouslySetInnerHTML={{
-      __html:
-        '<svg id="I52:3586;52:3570" layer-name="material-symbols:check-rounded" width="16" height="12" viewBox="0 0 16 12" fill="none" xmlns="http://www.w3.org/2000/svg" class="check-icon" style="width: 16px; height: 12px"> <path d="M5.95832 8.625L13.0208 1.5625C13.1875 1.39583 13.3819 1.3125 13.6042 1.3125C13.8264 1.3125 14.0208 1.39583 14.1875 1.5625C14.3542 1.72917 14.4375 1.92722 14.4375 2.15667C14.4375 2.38611 14.3542 2.58389 14.1875 2.75L6.54165 10.4167C6.37499 10.5833 6.18054 10.6667 5.95832 10.6667C5.7361 10.6667 5.54165 10.5833 5.37499 10.4167L1.79165 6.83333C1.62499 6.66667 1.54499 6.46889 1.55165 6.24C1.55832 6.01111 1.64527 5.81306 1.81249 5.64583C1.97971 5.47861 2.17777 5.39528 2.40665 5.39583C2.63554 5.39639 2.83332 5.47972 2.99999 5.64583L5.95832 8.625Z" fill="#5B5B5B"></path> </svg>',
-    }}
-  />
-);
